Move Definitions componenent to its own file
